fix: changed Vehicle class

+++ src/main/java/com/autonexo/vehicles/domain/models/commands/DeleteVehicleByIdCommand.java
@@ -1,5 +1,5 @@
 package com.autonexo.vehicles.domain.models.commands;
 
-public record DeleteVehicleByIdCommand(Integer vehicleId) {
+public record DeleteVehicleByIdCommand(Long vehicleId) {
 
 }
