+++ src/main/java/com/autonexo/vehicles/domain/models/commands/UpdateVehicleCommand.java
@@ -1,5 +1,5 @@
 package com.autonexo.vehicles.domain.models.commands;
 
-public record UpdateVehicleCommand(Integer vehicleId, Long mechanicId) {
+public record UpdateVehicleCommand(Long vehicleId, Long mechanicId) {
 
 }
