+++ src/main/java/com/autonexo/vehicles/domain/models/queries/GetVehicleByIdQuery.java
@@ -1,4 +1,4 @@
 package com.autonexo.vehicles.domain.models.queries;
 
-public record GetVehicleByIdQuery(Integer vehicleId) {
+public record GetVehicleByIdQuery(Long vehicleId) {
 }
