+++ src/main/java/com/autonexo/vehicles/interfaces/rest/resources/VehicleResource.java
@@ -1,4 +1,4 @@
 package com.autonexo.vehicles.interfaces.rest.resources;
 
-public record VehicleResource(Integer id, String plate, String make, String model, String year, Long driverId, Long mechanicId) {
+public record VehicleResource(Long id, String plate, String make, String model, String year, Long driverId, Long mechanicId) {
 }
